groq/requirements.txt
streamlit
langchain
langchain-groq
langchain-community
gtts
faiss-cpu
python-dotenv
